# openapi.yaml
openapi: 3.1.0
info:
  description: "This is the REST API documentation of the Spring Petclinic backend.\
    \ If authentication is enabled, use admin/admin when calling the APIs"
  termsOfService: https://github.com/spring-petclinic/petclinic-rest/blob/master/terms.txt
  title: REST Petclinic Backend API
  version: "1.0"
servers:
- description: Generated server url
  url: http://localhost
paths:
...
  /api/owners:
    get:
      description: "Lists owners, optionally filtered by search: a case-insensitive\
        \ fragment matched against the full name, address, city, telephone or any\
        \ pet name. Omitting search lists every owner."
      operationId: listOwners
      parameters:
      - description: Case-insensitive fragment matched against any listed column
        example: potter
        in: query
        name: search
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          content:
            application/json:
              examples:
                sample:
                  description: sample
                  value:
                  - id: 1
                    firstName: George
                    lastName: Franklin
                    address: 110 W. Liberty St.
                    city: Madison
                    telephone: "6085551023"
                    pets:
                    - id: 1
                      name: Leo
                      birthDate: 2010-09-07
                      type:
                        id: 1
                        name: cat
                      ownerId: 1
                      visits: []
                  - id: 2
                    firstName: Betty
                    lastName: Davis
                    address: 638 Cardinal Ave.
                    city: Sun Prairie
                    telephone: "6085551749"
                    pets: []
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: List owners
      tags:
      - owner-rest-controller
    post:
      operationId: addOwner
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerFieldsDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Create an owner
      tags:
      - owner-rest-controller
  /api/owners/count:
    get:
      operationId: countOwners
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: integer
                format: int64
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Count owners
      tags:
      - owner-rest-controller
  /api/owners/{ownerId}:
    delete:
      operationId: deleteOwner
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Delete an owner by ID
      tags:
      - owner-rest-controller
    get:
      operationId: getOwner
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OwnerDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Get an owner by ID
      tags:
      - owner-rest-controller
    put:
      operationId: updateOwner
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerFieldsDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Update an owner
      tags:
      - owner-rest-controller
  /api/owners/{ownerId}/pets:
    post:
      operationId: addPetToOwner
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetFieldsDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Add a pet to an owner
      tags:
      - owner-rest-controller
  /api/owners/{ownerId}/pets/{petId}:
    get:
      operationId: getOwnersPet
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: petId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Get a pet belonging to an owner
      tags:
      - owner-rest-controller
    put:
      operationId: updateOwnersPet
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: petId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetFieldsDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Update an owner's pet
      tags:
      - owner-rest-controller
  /api/owners/{ownerId}/pets/{petId}/visits:
    post:
      operationId: addVisitToOwner
      parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: petId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitFieldsDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      summary: Add a visit for an owner's pet
      tags:
      - owner-rest-controller
...
  /api/pets/{petId}:
    delete:
      operationId: deletePet
      parameters:
      - in: path
        name: petId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - pet-rest-controller
    get:
      operationId: getPet
      parameters:
      - in: path
        name: petId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - pet-rest-controller
    put:
      operationId: updatePet
      parameters:
      - in: path
        name: petId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - pet-rest-controller
...
components:
  schemas:
...
    OwnerDto:
      type: object
      properties:
        address:
          type: string
          description: The postal address of the pet owner.
          example: 110 W. Liberty St.
          maxLength: 255
          minLength: 1
        city:
          type: string
          description: The city of the pet owner.
          example: Madison
          maxLength: 80
          minLength: 1
        firstName:
          type: string
          description: The first name of the pet owner.
          example: George
          maxLength: 30
          minLength: 1
        id:
          type: integer
          format: int32
          description: The ID of the pet owner.
          example: 1
          minimum: 0
          readOnly: true
        lastName:
          type: string
          description: The last name of the pet owner.
          example: Franklin
          maxLength: 30
          minLength: 1
        pets:
          type: array
          description: The pets owned by this individual including any booked vet
            visits.
          items:
            $ref: "#/components/schemas/PetDto"
          readOnly: true
        telephone:
          type: string
          description: The telephone number of the pet owner.
          example: "6085551023"
          maxLength: 20
          minLength: 1
          pattern: "^[0-9]*$"
      required:
      - address
      - city
      - firstName
      - lastName
      - pets
      - telephone
    OwnerFieldsDto:
      type: object
      properties:
        address:
          type: string
          description: The postal address of the pet owner.
          example: 110 W. Liberty St.
          maxLength: 255
          minLength: 1
        city:
          type: string
          description: The city of the pet owner.
          example: Madison
          maxLength: 80
          minLength: 1
        firstName:
          type: string
          description: The first name of the pet owner.
          example: George
          maxLength: 30
          minLength: 1
        lastName:
          type: string
          description: The last name of the pet owner.
          example: Franklin
          maxLength: 30
          minLength: 1
        telephone:
          type: string
          description: The telephone number of the pet owner.
          example: "6085551023"
          maxLength: 20
          minLength: 1
          pattern: "^[0-9]*$"
      required:
      - address
      - city
      - firstName
      - lastName
      - telephone
    PetDto:
      type: object
      properties:
        birthDate:
          type: string
          format: date
          example: 2010-09-07
        id:
          type: integer
          format: int32
          description: The ID of the pet.
          example: 1
          minimum: 0
          readOnly: true
        name:
          type: string
          description: The name of the pet.
          example: Leo
          maxLength: 30
          minLength: 0
        ownerId:
          type: integer
          format: int32
          description: The ID of the pet's owner.
          example: 1
          minimum: 0
          readOnly: true
        type:
          $ref: "#/components/schemas/PetTypeDto"
        visits:
          type: array
          description: Vet visit bookings for this pet.
          items:
            $ref: "#/components/schemas/VisitDto"
          readOnly: true
      required:
      - birthDate
      - id
      - name
      - type
      - visits
    PetFieldsDto:
      type: object
      properties:
        birthDate:
          type: string
          format: date
          description: The date of birth of the pet.
          example: 2010-09-07
        name:
          type: string
          description: The name of the pet.
          example: Leo
          maxLength: 30
          minLength: 0
        type:
          $ref: "#/components/schemas/PetTypeDto"
      required:
      - birthDate
      - name
      - type
    PetTypeDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The ID of the pet type.
          example: 1
          minimum: 0
        name:
          type: string
          description: The name of the pet type.
          example: cat
          maxLength: 80
          minLength: 1
      required:
      - id
      - name
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
    VisitDto:
      type: object
      properties:
        date:
          type: string
          format: date
          description: The date of the visit.
          example: 2013-01-01
        description:
          type: string
          description: The description for the visit.
          example: rabies shot
          maxLength: 255
          minLength: 1
        id:
          type: integer
          format: int32
          description: The ID of the visit.
          example: 1
          minimum: 0
          readOnly: true
        ownerFirstName:
          type: string
          description: First name of the owner (server-populated).
          readOnly: true
        ownerId:
          type: integer
          format: int32
          description: ID of the owner of the pet (server-populated).
          readOnly: true
        ownerLastName:
          type: string
          description: Last name of the owner (server-populated).
          readOnly: true
        petId:
          type: integer
          format: int32
          description: The ID of the pet.
          example: 1
          minimum: 0
        petName:
          type: string
          description: Name of the pet (server-populated).
          readOnly: true
      required:
      - description
      - id
      - petId
    VisitFieldsDto:
      type: object
      properties:
        date:
          type: string
          format: date
          description: The date of the visit.
          example: 2013-01-01
        description:
          type: string
          description: The description for the visit.
          example: rabies shot
          maxLength: 255
          minLength: 1
      required:
      - description